// src/types.ts
export interface Dictionary<TValue> {
  [id: string]: TValue;
}
export type AtLeast<T, K extends keyof T> = Partial<T> & Pick<T, K>;

export interface ServiceCall {
  service: string;
  entity_id?: string;
  service_data?: Dictionary<any>;
}

export interface Action {
  name?: string;
  service: string;
  service_data?: Dictionary<any>;
  icon?: string;
  variables?: VariableDictionary;
  supported_feature?: number;
}
// ...
export interface GroupConfig {
  name: string;
  icon?: string;
  include?: string[];
  exclude?: string[];
}

/* entities */

export interface EntityElement {
  id: string;
  name: string;
  icon?: string;
  exclude_actions?: string[];
}

export interface EntityConfig {
  name?: string;
  icon?: string;
  actions?: Action[];
  exclude_actions?: string[];
  states?: string[] | { min: number; max: number; step?: number; unit?: string };
}
// ...
export interface CardConfig extends LovelaceCardConfig {
  discover_existing: boolean;
  standard_configuration: boolean;
  title: boolean | string;
  time_step: number;
  show_header_toggle: boolean;
  show_add_button?: boolean;
  display_options: {
    primary_info: string[] | string;
    secondary_info: string[] | string;
    icon: string;
  };
  include: string[];
  exclude: string[];
  groups: GroupConfig[];
  customize: Dictionary<EntityConfig>;
  tags?: string[] | string;
  exclude_tags?: string[] | string;
  sort_by: string[] | string;
}